# derp: Pool byte allocs