Previously, bytes were being allocated on every send, which could be costly. Instead, allocating
bytes from a pool would reduce the memory overhead. At very tiny msg sizes (10 bytes),
allocating for each works better.

I would also note the `ns/op` timings below are probably not correct, since there is a ton of
noise on my machine, but if those can be trusted it is slightly faster.

```
name                      old time/op    new time/op    delta
SendRecv/msgsize=10-8       7.92µs ±26%    7.56µs ±24%     ~     (p=0.124 n=20+16)
SendRecv/msgsize=100-8      8.29µs ±24%    8.31µs ±17%     ~     (p=0.789 n=20+16)
SendRecv/msgsize=1000-8     9.30µs ±22%    8.80µs ±22%     ~     (p=0.087 n=19+17)
SendRecv/msgsize=10000-8    24.9µs ±23%    23.1µs ±19%   -7.09%  (p=0.025 n=19+17)

name                      old speed      new speed      delta
SendRecv/msgsize=10-8     1.25MB/s ±16%  1.29MB/s ±25%     ~     (p=0.233 n=19+16)
SendRecv/msgsize=100-8    12.2MB/s ±30%  12.1MB/s ±20%     ~     (p=0.783 n=20+16)
SendRecv/msgsize=1000-8    108MB/s ±20%   113MB/s ±15%     ~     (p=0.142 n=19+16)
SendRecv/msgsize=10000-8   399MB/s ±16%   436MB/s ±18%   +9.39%  (p=0.008 n=18+17)

name                      old alloc/op   new alloc/op   delta
SendRecv/msgsize=10-8         112B ± 0%       96B ± 0%  -14.29%  (p=0.000 n=19+16)
SendRecv/msgsize=100-8        208B ± 0%       99B ±10%  -52.61%  (p=0.000 n=20+16)
SendRecv/msgsize=1000-8     1.12kB ± 0%    0.10kB ± 0%  -91.43%  (p=0.000 n=18+15)
SendRecv/msgsize=10000-8    19.8kB ± 1%    10.4kB ± 0%  -47.44%  (p=0.000 n=19+17)

name                      old allocs/op  new allocs/op  delta
SendRecv/msgsize=10-8         3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=17+17)
SendRecv/msgsize=100-8        3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=20+17)
SendRecv/msgsize=1000-8       3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=18+17)
SendRecv/msgsize=10000-8      3.00 ± 0%      3.00 ± 0%     ~     (all equal)
```