@bradfitz 2021-06-10
You need more than 1 run to get good benchstat output. (Notice the `~` and poor `p=1.000` value)